+++ src/cpath.c
@@ -4,9 +4,9 @@
 
 void cpath_init(struct cpath* path, struct dp_state_table const* state_tbl, struct dp_trans_table const* trans_tbl)
 {
-    path->nstates = dp_state_table_nstates(state_tbl);
-    path->bits_state = malloc(sizeof(*path->bits_state) * (path->nstates + 1));
-    path->bits_trans = malloc(sizeof(*path->bits_trans) * path->nstates);
+    path->nstates = (uint16_t)dp_state_table_nstates(state_tbl);
+    path->bits_state = malloc(sizeof(*path->bits_state) * (unsigned)(path->nstates + 1));
+    path->bits_trans = malloc(sizeof(*path->bits_trans) * (unsigned)path->nstates);
     path->bits_state[0] = 0;
 
     for (uint_fast16_t tgt = 0; tgt < path->nstates; ++tgt) {
@@ -20,10 +20,10 @@ void cpath_init(struct cpath* path, struct dp_state_table const* state_tbl, stru
             depth = (uint_fast8_t)MAX(max_seq - min_seq, depth);
         }
         path->bits_state[tgt + 1] = path->bits_state[tgt];
-        path->bits_trans[tgt] = (uint8_t)bits_width(dp_trans_table_ntrans(trans_tbl, tgt));
+        path->bits_trans[tgt] = (uint8_t)bits_width((uint32_t)dp_trans_table_ntrans(trans_tbl, tgt));
         path->bits_state[tgt + 1] += path->bits_trans[tgt];
         /* Additional bit (if necessary) for invalid transition or seq_len */
-        path->bits_state[tgt + 1] += bits_width(depth + 1);
+        path->bits_state[tgt + 1] += bits_width((uint32_t)((unsigned)depth + 1));
     }
     path->bitarr = NULL;
 }
@@ -36,7 +36,7 @@ void cpath_deinit(struct cpath const* path)
         free(path->bitarr);
 }
 
-void cpath_setup(struct cpath* path, uint32_t len)
+void cpath_setup(struct cpath* path, uint_fast32_t len)
 {
     uint64_t size = (uint64_t)path->bits_state[path->nstates] * (uint64_t)len;
     path->bitarr = bitarr_realloc(path->bitarr, size);
